backend/requirements.txt
# Backend Requirements - Production

# Django and DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Authentication
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# API Documentation
drf-spectacular==0.27.0

# Environment Variables
python-decouple==3.8

# Utilities
pytz==2023.3
requests==2.31.0

